Moved factories to new files
