+++ eastwood/protocols/base_protocol.py
@@ -4,8 +4,7 @@ Twisted protocol and factory for TCP communication
 
 import logging
 from quarry.net.protocol import BufferUnderrun
-from quarry.types.buffer import buff_types
-from twisted.internet.protocol import Factory, Protocol
+from twisted.internet.protocol import Protocol
 
 class PacketDispatcher:
 	"""
@@ -141,28 +140,3 @@ class BaseProtocol(Protocol, PacketDispatcher):
 		Returns:
 			id: id of the packet
 		"""
-
-class BaseFactory(Factory):
-	"""
-	Base Factory, contains some common variables and passes needed arguments for BaseProtocol
-	"""
-	protocol=BaseProtocol
-
-	def __init__(self, protocol_version, handle_direction):
-		"""
-		Args:
-			protocol_version: minecraft protocol specification to use
-			handle_direction: direction packets being handled by this protocol are going (can be "clientbound" or "serverbound")
-		"""
-		self.protocol_version = protocol_version
-		self.buff_class = self.get_buff_class()
-		self.handle_direction = handle_direction
-		self.other_factory = None # Other factory is assigned by hand to prevent chicken egg problem
-
-	def get_buff_class(self):
-		for version, buff_class in reversed(buff_types):
-			if self.protocol_version >= version:
-				return buff_class
-
-	def buildProtocol(self, addr):
-		return self.protocol(self, self.buff_class, self.handle_direction, self.other_factory)
